+++ src/scripts/plugins/toast.js
@@ -7,14 +7,20 @@ const UI_TOAST = {
   timeoutMs: {
     MIN: 2000,
     MAX: 3500,
-    DEFAULTS: 2000
+    DEFAULTS: 2750
+  },
+  position: {
+    TOP: 'top',
+    BOTTOM: 'bottom',
+    CENTER: 'center'
   }
 };
 
 const DEFAULT_OPTIONS = {
   className: '',
   timeoutMs: UI_TOAST.timeoutMs.DEFAULTS,
-  message: ''
+  message: '',
+  position: UI_TOAST.position.BOTTOM
 };
 
 let globalOptions = DEFAULT_OPTIONS;
@@ -43,10 +49,16 @@ function createToast(options) {
       };
     },
     computed: {
+      positionClassName() {
+        return ['top', 'center'].includes(this.options.position)
+          ? `mdc-toast--${this.options.position}`
+          : '';
+      },
       className() {
         return [
           'mdc-snackbar',
           'mdc-toast',
+          this.positionClassName,
           this.options.className,
           {
             'mdc-snackbar--opening': this.opening,